Analyse mixture results in a vectorised fashion

Refer to analyse_mixture_result() for details. Essentially, analyse_mixture_result() is run on each element of mixture_results.

analyse_mixture_results(
  mixture_results,
  unique_haps_in_mixture_list,
  unique_haps_in_mixture_counts_list
)

Arguments

mixture_results

List of n mixture results from mixture_info_by_individuals_2pers(), mixture_info_by_individuals_3pers(), mixture_info_by_individuals_4pers(), mixture_info_by_individuals_5pers()

unique_haps_in_mixture_list

List of n included unique haplotypes, one for each element in mix_res

unique_haps_in_mixture_counts_list

List of n population counts of the included unique haplotypes

Value

A list with lists of numeric quantities

Details

NOTE: Only takes up to 9 contributors!