Populate 1-locus autosomal DNA profile in pedigrees with infinite alleles mutation model.

Populate 1-locus autosomal DNA profile from founder and down in all pedigrees. Note, that all founders have type 0 to begin with.

population_populate_autosomal_infinite_alleles(
  population,
  mutation_rate,
  progress = TRUE
)

Arguments

population

Population in which to populate genotypes

mutation_rate

Mutation rate between 0 and 1 (both included)

progress

Show progress

Details

The maternal allele is taken by random from the 2*N[g] alleles in the previous generation consisting of N[g] males with descendants in the live population.

This is also why this is not using pedigrees but instead the population.

Note, that pedigrees need not be inferred.

See also