Mutational information about Y-STR markers

A dataset from yhrd.org (and their sources) containing mutational information about Y chromosomal short tandem repeat (Y-STR) markers used in forensic genetics.

ystr_markers

Format

A data frame with 29 rows and 5 variables:

Marker

name of Y-STR marker

Meioses

number of meioses observed

Mutations

number of mutations observed in the corresponding number of Meioses

MutProb

point estimate of mutation probability, MutProb = Mutations/Meioses

Alleles

observed alleles

Source

http://www.yhrd.org

Details

Note, that loci with duplications (DYS385a/b as well as DYF387S1a/b have been split into two loci).